java/src/main/java/com/yourtion/leetcode/medium/c0073/Solution.java
package com.yourtion.leetcode.medium.c0073;

import java.util.Arrays;

/**
 * 73. 矩阵置零
 *
 * @author Yourtion
 * @link https://leetcode-cn.com/problems/set-matrix-zeroes/
 */
public class Solution {
    public void setZeroes(int[][] matrix) {
        boolean isCol = false;

        for (int i = 0; i < matrix.length; i++) {
            if (matrix[i][0] == 0) {
                isCol = true;
            }
            for (int j = 1; j < matrix[i].length; j++) {
                if (matrix[i][j] == 0) {
                    matrix[i][0] = 0;
                    matrix[0][j] = 0;
                }
            }
        }
        for (int i = 1; i < matrix.length; i++) {
            for (int j = 1; j < matrix[i].length; j++) {
                if (matrix[i][0] == 0 || matrix[0][j] == 0) {
                    matrix[i][j] = 0;
                }
            }
        }
        if (matrix[0][0] == 0) {
            Arrays.fill(matrix[0], 0);
        }
        if (isCol) {
            for (int i = 0; i < matrix.length; i++) {
                matrix[i][0] = 0;
            }
        }
    }
}
